medical_chatbot/requirements.txt
google-generativeai
python-dotenv
langchain
langchain-google-genai
langchain-community
langchain-core
chromadb
langchain-chroma
pandas
numpy
tqdm
tavily-python
gdown
langchain-openai
langchain-huggingface
openai
transformers
faiss-cpu
gradio
jedi
